\docType{data}
\name{MarineVessels}
\alias{MarineVessels}
\title{
Basic information for a large portion of the world's active marine vessels
}
\description{
A dataset containing marine vessel names and selected metadata. 
}
\usage{
MarineVessels
}
\format{
A data frame with x observations on y variables. This data frame contains the following columns:
  \describe{
    \item{\code{}}{}
  }
}
\source{

}
\examples{
head(MarineVessels)
}
